+++ Toris/Assets/Scripts/UIToolkit/ScriptableObjects/InventoryContainerSO.cs
@@ -8,5 +8,8 @@ namespace OutlandHaven.Inventory
     {
         public int SlotCount = 20;
         public ScreenType AssociatedView = ScreenType.None;
+
+        [Header("Optional. Predefined filters for specific slots by index.")]
+        public SlotFilterType[] PredefinedFilters;
     }
 }